requirements.txt
asgiref==3.5.2
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
mysqlclient==2.1.1
PyJWT==2.5.0
pytz==2022.4
sqlparse==0.4.3
tzdata==2022.4